Parser.Stream: replace xml-conduit with xeno (experimental hack)
Xeno only supports strict ByteStrings as input. But the SAX parser
only does the lightest of validation at a very 'local'
level (e.g. closing tag at end of a full attribute list). This commit
intitiates a sax parse of multiple fragments fed via a
ByteString source. The ByteString can't be split arbitrarily, but
needs to be split in a way that doesn't cause a parse
failure[1]. This hence relies on incidental details of how the xeno
parser chooses to index into the string.

We thus split each bytestring in a very hacky way: taking the largest
fragment before the final '<' character in the stream, and then
prepending the remainder onto the next yielded fragment. This is
certainly a bad idea; any comment or text value would/could contain that
character. (I also tried a 'space' rather than a '<', but this still
caused errors).

It does seem to work for our test xlsx files, but I'm not sure this is
a robust enough option and will next try expat and/or libxml2's sax
parser.

And most bizarrely, this parser appears to be about 2x slower than
xml-conduit. For a 150k-row sheet, countRowsInSheet takes 5.9 seconds
with xml-conduit, and 12.26 with this version via xeno. (Ran from
GHCi, with '-fobject-code -O2'; both figures are the mean of 4
runs). This is 2.1x slower. I also tested this with supercede
proprietary processor (hence the library was compiled via cabal rather
than GHCi), and found an overall 1.6x slowdown (the difference
presumably from more work spent in identical code downstream from this
package.)

One reason for the slowdown could be that we only use xeno in the most
minimal way possible (decoding the bytestrings as utf8 into sax events
defined in the 'xml-types' package, rather than using ). I guess we
could be more aggressive about sticking to bytestrings and avoiding
the intermediate xml-types event structures, but we do need some
intermediate type before SheetItem, and we ultimately need to decode
utf8 anyway. Still, this is worth trying quickly as the parser doesn't
need to utf8-decode attribute keys (and we don't use most of the values).

[1]: an alternate design, in theory, could split or join only _after_
getting a parse error, leaving some things up to chance. A sketch: 1)
parse and issue sax events until an error occurs, 2) catch that
error along with the leftover text, 3) prepend that text to the
next yielded ByteString, _and/or_, split that text into two
pieces, try parsing the first piece again (and then repeatedly divide
on failure); 4) prepend the latter piece to the next yielded chunk, and so
on. But I think this has the potential to get stuck quickly: if the
next ByteString also ends at a place where xeno will throw an error,
then a larger ByteString will keep might keep getting built up. (Or if
we split arbitrarily in 2, then I think the chance of the string being
split at a parseable spot is _very_ low, and we might get stuck with
single charaters)