Stream parser: Add callback-based expat interface
The approach where we ask the user for an IO callback rather than try
to integrate the xml library with conduits has been
fruitful. With a release build, we have the following results:

The command:
  time result/bin/countrows prem-est-150k.xlsx \
    <LIB> +RTS -N4 -M7G -A4M -s -RTS

when run with <LIB> as the following:

UseXmlConduit: 10.7s, max 167MB residency
UseLibxml:     7.1s,  max 46MB residency
UseHexpat      3.9s,  max 46MB residency

The UseLibxml result was promising enough that trying to use a
callback-approach with expat made sense.

However Hexpat exposes only a somewhat tricky ListT class that requires the
same ListT instance for the input as for the output. That's not an
option here as we need to use conduits for the input
bytestring (supplied to us via the "zip" library), but avoid conduits
at the consumer level.

But we can still use the lower-level hexpat API provided we expose one
extra function in the hexpat library (copy and pasted here in the
HexpatInternal module, conscious that this isn't the ideal solution).

So we've done that and we find a great speedup compared both to
xml-conduit and libxml.