Add row properties
This allows users to set rows to a particular height.

This also includes a small refactor:
The primary reason for this is to expose the reader monad
within the row.
However,
this may give a speedup as well because we're relying more
on conduit and less on (slow) list concatenation.
Furthermore it looks a lot nicer because the implementations
no longer have to manually construct the Event types and get
surounded with tags trough the callback.