Add style parsing and writing
Also modifies the test to confirm this works.
The styles are fully unstructured,
which should be good enough for our usecase.
It's still possible to use the existing formatted cell
API with this because it ends up being the same
unstructured type in xlsx.

\squash log

Add reading and writing of styles

Add styles to content types

Add rel for theme

Fix relationship bug as the validitor program said it was

Remove uneccisary newlines