Improve getWorkbookInfo API to include name and r:id attribute
The "r:id" attribute is the number that corresponds to the
xl/worksheets/sheetN.xml files, not the "sheetid" attribute. The
previous code conflated these two values and meant getWorkbookInfo
could not be used for the number to pass to readSheet.

Changes the sheets representation to a list, as I don't think it's
right to prioritise sheetid vs r:id as a key into a map. It also
better reflects the XML.

Purposefully doesn't change the 'readSheet' function for
backwards-compat, but it does change the getWorkbookInfo function,
which is a breaking change although I don't think the previous version
would have been very useful.

Adds a readSheetByName helper function.