Revert "Fix confusion between sheet's r:id and sheetId"
This reverts commit 9f4d480f4d5c53eb28a012d4aaecd2274ef60f55.

* In practice, Excel writes the xlsx zip package such that the <N> in
xl/worksheets/sheet<N>.xml always corresponds to the relationship
ID (aka the sheet element's r:id attribute, and not its sheetId
attribute).

* But the xlsx spec doesn't define any required correspondence. The <N>
in xl/worksheets/sheet<N>.xml doesn't necessarily correspond to either
the r:id or the sheetId. The relationships file has to be used to get
the full path name, and sheet file names don't even have to be follow
the sheet<N>.xml pattern . MS have an example of this case here.

* The stream writer doesn't follow the Excel app's convention, but it
does follow the spec in not relying on a correspondence.

Therefore to ensure this works consistently, we need to parse the
relationships file and not rely on the naming convention of
sheet<N>.xml, which is not even part of the spec.