Add 'readPartialSheet' for quick sheet sampling
Regardless of how many rows the caller wants to process, `readSheet`
will end up traversing and allocating a `SheetItem` for each row in the
sheet. This makes traversing just the start of a sheet, which we want
to do to get a sample of all the rows in a worksheet, very inefficient.

This adds a new function, `readPartialSheet`, which behaves similarly to
`readSheet` but which will not allocate more rows than necessary when
traversing a worksheet. The user can decide if they want to continue
receiving rows using the `Continue` datatype when consuming rows - once
the user has processed enough rows, they can return `Stop` to exit sheet
processing early.

This is a bit of a leaky abstraction - ideally, the user would be able
to consume a `ConduitT` of sheet items, which is automatically lazy in
allocation (since it is driven by downstream demand), and more flexible.
But I first want to try to understand why `ConduitT`s are not used for
this purpose already, before changing the design.